Updated templating and added hover effects
I updated the templating to use the one demonstrated in class instead
our manual templating.  This takes care of a lot of the stuff for us and
makes our code cleaner.

I also renamed the template files to .html so that they can be
(somewhat) viewable in browsers while designing.  This also helps with
code formatting for apps like Notepad++ that format based on file
extension.

Lastly I added a hover over effect for our links.